requirements.txt
black==24.4.2
click==8.1.7
colormath==3.0.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
tomli==2.0.1
typing_extensions==4.12.0
